+++ client/src/mutations/useLoginMutation/api.ts
@@ -1,0 +1,7 @@
+import { api } from '@api/index';
+import type { AuthFormInput, Login } from './types';
+
+const RESOURCE = '/users';
+
+export const login = async (data: AuthFormInput) =>
+  api.post<Login>(`${RESOURCE}/login`, data);
